move _timeDeltaToHumanReadable to buildbot.util.human_readable_delta
* there's an overlap between human_readable_delta and formatInterval
